test: cover analytics aggregation against fixtures
